requirements.txt
# === Core Framework ===
fastapi
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
uvicorn
python-multipart
websockets

# === LangChain & AI ===
langchain-google-genai
langchain-mcp-adapters
langgraph
langchain-core

# === Authentication & Security ===
bcrypt
python-jose[cryptography]

# === Database ===
supabase

# === HTTP & Network ===
requests
httpx
aiohttp>=3.9.0

# === Utils ===
python-dotenv
python-dateutil>=2.8.2
sentence-transformers
numpy
PyPDF2
pillow
pytesseract

# === PAIA Protocol Dependencies ===
pydantic>=2.5.0
jsonschema>=4.20.0